💡 (Refresh token) Add TODO for what to do with refresh token
The refresh token has to be added to the blocklist _at some point_ so that it can't be re-used forever. We can add it after the first time it's used so users have 1x fresh token, 1x non-fresh token usage, before they need to re-authenticate.

We should mention at the point of implementing that the refresh token can be added to the blocklist at any time; it's up to the developer.